Allows the modification of default fields
This addition in getFieldComponent function allows us to override default type fields components by customized fields without needing to specify "ui:field" in uiSchema. This is useful if you want to render the schema in another context than a form. I have added an exception for "object" type because of the complexity of the component.