Improved documentation on form initialization
Through the docs I got the impression that `formData` prop is used for initialization only, which led to a hard-to-detect bug that was happening when the parent element was re-rendering. So I decided to update the documentation so it's more clear. Please feel free to modify the texts that I wrote so they fit well with the rest of the documentation