Allows injecting custom jsonschema Validator instance
The current custom validation does not have access to the schema
in the same way that a jsonschema validator does, by allowing
user to specify the validator many validation tasks become much
simpler.

This is related to #79 which was closed when the custom field validation was implemented,
however that is not as powerful as a jsonschema validator can be.