Fix behaviour of arrays with minItems and default
Fix bug whereby defaults in an array were munged together and added to extra item's formData.
Behaviour now is to fill the array with the object level default array, and then add extra items up to minItems.
These extra items will be either empty or the item-level default.
Added demonstration of this in the Arrays tab of the playground.