Allow schema $refs to not have to start with `#/definitions`
We're using this module to render forms for Swagger/OAS files. $refs
in swagger are stored in `components/schemas` at the top level of the object.

This change allows $ref strings to not always start with `#/definitions/`
which is a much easier solution than us having to recursively modify $refs
to prefix with this value.

All existing tests are passing, with a new test for this functionality.

The change was to make the definitions part of the regular expression an
optional non-capturing group.