[V2] Upgrade to react 16 (#1408)
* upgrade to react 16

* bump react peer dependency to 16

* remove react-lifecycles polyfill

* unskip react 16 test

* Fix array tests, fixes #1427

componentWillReceiveProps is not called after setState, but
getDerivedStateFromProps is called after setState. This caused the array
key tests to fail, because when onAddIndexClick is called, it sets the
state first to update keyedFormData. Only afterwards is the onChange
handler called with the new form data. And, getDerivedStateFromProps is
called in between, so it ends up being called with the old form data
instead of the new one.

Used the technique in https://stackoverflow.com/a/57208053 to work
around this, by not calculating derived state in
getDerivedStateFromProps when keyed form data is
updated.

* Remove console.error stub to fix tests

* Fix getWidget behavior for forwarded refs

* Replace componentWillReceiveProps with UNSAFE_componentWillReceiveProps