Better consistency with Material UI patterns with usage of <TextField /> (#1790)
This PR brings better consistency with Material UI patterns. It attempts to use < TextField /> from Material UI whenever possible. There is an extra <FormControl /> component in the current component tree and this fixes that (see images for before/after of <TextWidget />). It also moves the required property up to the main <FieldTemplate /> and then sets it on the specific <TextField /> components as well. This change makes the error behavior more consistent with the way <TextField /> usually works out of the box. I also made minor adjustments to the <List /> and <ListItem /> used for errors to make the UI more consistent with what you see at https://material-ui.com/components/text-fields/#validation. Finally it adjusts widgets like <SelectWidget /> to use <TextField /> rather than reconstruct that component.