additionalProperties: retain object title when renaming key (#1751)
Fixes #1749. This reverts part of #1401 (the conditional setting of the title in the wasPropertyKeyModified logic in ObjectField.js) -- this piece of logic doesn't seem to contribute to the fix in #1401, and it's what caused #1749.

Additionally, the test that was added in #1401 didn't actually check to make sure the issue it was solving was resolved; that test passed even before #1401 was merged. The issue that #1401 was that onChange was called with the new key, but the new key did not show up in the DOM. I've updated that test to check for this behavior (updating the new key in the DOM) now.