[GH-2090] [fluent-ui] fix: Text inputs render with invalid attribute
TextWidget generated `<input type="string" ... >`
HTML which is invalid as per
https://html.spec.whatwg.org/multipage/input.html#states-of-the-type-attribute

`<input type="text" ... >` would be apropriate.

The generation of invalid HTML results in undefined behaviour.
I got on the trail of the issue, because
[testing-library/user-event](https://github.com/testing-library/user-event#typeelement-text-options)
acted in strange ways.

This PR results in the generation of valid HTML and updates the tests.

fix #2090