Don't render label twice with implicit "object" schema
`getSchemaType` is clever enough to figure out that a schema with
`properties` or `additionalProperties` has `type` "object", even if the
type is not explicitly defined, but `getDisplayLabel` does not have the
same logic, which results in displaying the label twice in cases where
the schema was guessed to be object.

Using `getSchemaType` as well in `getDisplayLabel` bring consistent
behavior and the label is rendered only once in that case.

Fixes #2501