Can now specify idSeparator to control how id is created
By default, '_' is used as a separator for the different parts that
make up a field's id. This can be problematic when '_' is part of a
field's name.

This commit adds the ability to specify idSeparator when building a
form which allows users use a character that isn't used in field
names.

With the idSeparator option set to '.', we would get:

'root.field_a.field_123' instead of 'root_field_a_field_123'.