docs: remove reference to the deprecated uiSchema `classNames` property
In the v5 migration guide, it is explained that `classNames` property within
uiSchema is now deprecated in favor of `ui:classNames` property. In developer
use the legacy classNames, a deprecation warning is displayed within the console
. But currently, in the uiSchema section within the quickstart guide,
documentation still refer to this deprecated property that is weird for
newcomers which will bootstrap a project with deprecation warnings.