Update LICENSE.md (#4174)
* Update LICENSE.md

Dear rsjf-team, I have a request. We are using the great rjsf packages to build a forms solution for a company. This company requires a copyright notice in the license file. Could you please replace the example copyright notice by a copyright of your team. I know it is strange at that location, but it would solve our problems.

Examples where they added names at that location in the license file:

- https://github.com/eisberg-labs/actix-actor-expect/blob/c28e751d219535c3d4d2b5579ddd281440e5ea87/LICENSE-APACHE#L4
- https://github.com/Swiftify-Corp/Swift-30-Projects/blob/5140f35ca0cb3767807bc518084809030a77139a/License.md?plain=1#L4

* Update LICENSE.md

- Added the start date from when the repo was created

---------

Co-authored-by: Heath C <51679588+heath-freenome@users.noreply.github.com>