Freeze docs for v5
Use `npm run docusaurus docs:version 5.24.9` to freeze docs for v5.24.9. Update the label for these frozen docs to 'v5'. Update the label for the current docs to indicate the current version. Update tsconfig to allow importing package.json to resolve the package version.

Any changes to v5 docs after this is merged must be made to the versioned copy.